refactor: make ParseProfileRef a test-only helper
ParseProfileRef only ever reconstructs a ProfileRef from its URN to verify
String() round-trips; no production code parses refs back from URNs (they are
built from request inputs). Move it into a _test.go file so it cannot become a
load-bearing part of the package's public API, and strengthen the
pipeline-scope test to assert the full parsed struct rather than just an empty
ScopedRepository.
